+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVPYAD9HXY8X2VFCTR70VX6N"
+        "01KVQHZSVETTGZF160N9SFX3XA"
     )
     r"""The unique identifier of the data source"""