+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVPYAD9EY2Q65229HDN462QN"
+        "01KVQHZSV9RJGXCABK37ZHBBKG"
     )
     r"""The unique identifier of the data source"""