+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVPYAD9G8HVRNJAS98S08AEE"
+        "01KVQHZSVCWMTS7891JQ6RHRNS"
     )
     r"""The unique identifier of the data source"""