+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -1960,7 +1960,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KVPYACWBFEP874391SM9X8JK"
+    id: Optional[str] = "01KVQHZSDHD8CQK5X14SBQ1TW8"
 
     description: Optional[str] = None