+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVPYAD9M0XWTZCFGSB8KDC26"
+        "01KVQHZSVK0FQZ9X8PN6TK9R54"
     )
     r"""The unique identifier of the data source"""